Title: 나선형 삼각형, Time: 102ms, Memory: 8MB, Status: Passed - CodeTree

+++ 240802/나선형 삼각형/spira-triangle.java
@@ -12,7 +12,7 @@ public class Main {
 	public static void main(String[] args) throws IOException{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
-		int[] dp = new int[n+1];
+		long[] dp = new long[n+1];
 		
 		if(n<=3) {
 			System.out.println(1);
